0.1.9 Release

+++ lib/gpx/src/addWaypoint.cpp
@@ -2,13 +2,13 @@
  * @file addWaypoint.cpp
  * @author Jordi Gauchía (jgauchia@gmx.es)
  * @brief  Add Waypoint functions
- * @version 0.1.8
- * @date 2024-11
+ * @version 0.1.9
+ * @date 2024-12
  */
 
 #include "addWaypoint.hpp"
 
-extern const int SD_CS;
+extern const uint8_t SD_CS;
 
 wayPoint addWpt = {0, 0, 0, (char *)"", (char *)"", (char *)"", (char *)"", (char *)"", (char *)"", 0, 0, 0, 0};
 wayPoint loadWpt = {0, 0, 0, (char *)"", (char *)"", (char *)"", (char *)"", (char *)"", (char *)"", 0, 0, 0, 0};
